Make code snippets executable Python code
This adds elements to each slide to allow the execution of the code
shown. Also moves all of the code shown in the examples, questions, and
answers into these blocks so that students can run the code and play
with the results. Should make it a bit more interactive and fun. Not to
mention should make it easier for those unable to make it to explore the
same concepts.