Reintroduce the weighted mean with Dask Arrays
Have the students revisit implementing the weighted mean with Dask
Arrays. This let's them get an idea of how Dask Arrays can aid in
solving this problem under the hood. Also with the foreknowledge of the
primitives that Dask Arrays are building off of, the students should be
able to realize what Dask Array is doing here is nothing more than what
they did earlier.